(feature) InspectionsFinder
Responsible for querying the WordPress database and wrapping the results
in an object to make it easier to get at the ACF data and permalink

The search is for 'plugin links which include the slug, surrounded by
forward slashes. This is not perfect, but will find plugins like:

  - https://wordpress.org/plugins/google-tag-manager/
  - https://premium.wpmudev.org/project/new-blog-template/

And will mean that if the slug is

    advanced-custom-fields

then the search will return inspections for

    https://wordpress.org/plugins/advanced-custom-fields/

but not

    https://wordpress.org/plugins/advanced-custom-fields-table-field/

NOTE: The tests for this are an integration test for the Inspection
class. I think the way to decouple these would be to use a Factory
object to create the inspections an inject that in the constructor of
the InspectionsFinder. That feels a bit too obtuse for the moment.

NOTE: allowMockingNonExistentMethods appears to be incompatible with
WP_Mock, so let's selectively disable it. Raised an issue with the
project about this: <https://github.com/10up/wp_mock/issues/77>