(chore) Replace Scopes with beforeEach
Peridot provides a mechanism called 'scopes' for mixing in test helper
functions: https://peridot-php.github.io/scopes.html

However this seems intended for shared helpers across multiple tests,
and not for helper functions which are only relevant to one test file.

PSR-2 doesn't allow classes to be defined in the same file as other
code, so I've switched to defining functions in a beforeEach call.

Also had to redefine FakeParams as a mock instead of a straightforward
class, (PSR-2 didn't accept the wordpress snake-cased function name, and
I couldn't see a way around it).

Also removed the 'Mockery::close()' call: despite what the docs say this
doesn't seem to be required.