(refactor) extract a Router
This refactoring is probably premature at this point. The point is to
separate out the building of the Controller class from the creation of
the API endpoints.

I don't feel like there's anything here which can be unit tested: it's
all configuration and implementation detail, so any test would break if
the implementation changed.

The only part which seems amenable to testing is the regex, but I can't
see how this could be tested without exercising the real
`register_rest_route` function.