(fix) include timezone in test
This test passes without this on all versions of PHP on CI, but
apparently fails on 7.0 on some local dev machines
(https://git.govpress.com/dxw-advisories/app/merge_requests/17/diffs#note_7314)

Presumably this is because we are now in British Summertime and so times
are up to an hour out?

No way I can see to test this, but the tests pass with this addition
locally.