(fix) Handle timezones which don't overlap UTC
(originally committed by Tom Adams)

The CI and production servers should always both be running in UTC.
However, our local development machines will not always be running in
GMT, so we need to ensure that the test suite is tolerant of being run
in different timezones.

To simulate this, we can set timezone in CI container to one which never
overlaps with UTC+0000 - in this case Pacific/Auckland. This led to a
test failure in the JSONInspectionsFinder spec, which wasn't specifying
a timezone when creating test data, and was using the local timezone
instead.

However, this oddly didn't fail when running the tests on my machine in
BST