(fix) Hide failure criteria when there are none
(similar to how we did it in the old theme)

The logic previously was:

  - Show the whole section if there were any failure criteria

That implied that in order for there to be a reason there HAD to be some
failure criteria checked, which is not the case e.g. if the review is
for some unusual reason, or simply at the reviewer's discretion.

The logic now is:

  - Display the section if the review was yellow or red
  - Display the list of selected failure criteria only if there are some
    checked

Additional changes;

  - Remove the table headings - they don't make sense: everything is a
    failure.
  - CSS change to add back in the top border on the table