(fix) Reason and Findings are required
These should really be on every review (except that 'no issues found'
means no reason). Note that the template expects them to be there: if
they aren't then we get titles on the page with no content.

I've tested that when the reasons field is blank but not shown (because
no issues were found) then the form still submits without errors.