(chore) remove notices
This section used to be where we recorded security incidents. It is the
only reason that we have new-members-only on the site (we were using it
to share details of attacks with Nacin and other external people).

These now get recorded on Bikeshed (https://bikeshed.dxw.com), and the
old ones have been migrated there, so we can remove this functionality
entirely.