(chore) Migrate FetchPluginDetails from old theme
This functionality relies on javascript, and as far as I can see that
script won't work in its current form in the new theme - and it wasn't
working in the old theme either. So I've migrated the class over to
allow us to work on fixing this functionality, but it's not
actually enabled anywhere at the moment.