Move FetchPluginDetails into new theme
In the process of migrating FetchPluginDetails to the new theme, I have
also done some significant refactoring. Some of this has simply been to
ensure the module works in the new theme (e.g. by referencing the
correct input IDs), but I have also rejigged the class structure to
improve separation of concerns and make the code more testable.

The old version pulled a "short description" from a the plugin's page on
https://wordpress.org/plugins/. The element this referenced no longer
appears to exist on plugin pages, so the description is now taken from
the first paragraph of the plugin description text.