(fix) Only autofill form if value is a WordPress plugin link or slug
Before we had an issue where if you pasted a non-WordPress plugin link
the script would treat the value as a slug and autofill the form causing
the data to be misformed.

Now it will only attempt to autofill the form if the value matches a
WordPress link or slug, all other values will be ignored.