Migrate from npm to Yarn, remove bower
I've had to upgrade grunt-sass in order to do this, as the old version would cause an error when trying to install. That means I've also had to add node-sass, as you have to provide a sass implementation option for newer versions of grunt-sass.

We don't use bower anymore, as it's simpler to just do everything in yarn. So I've also got rid of the bower things and loaded jQuery via yarn as well.