Apply new brand
This is a very rough pass to apply the brand. It sorts out the logo, the
fonts, and most of the colours.

The red/orange/green colours used for plugin ratings don't really exist
in our brand palette, so I've left those alone.

The content and design of this site could use a *lot* of attention.
Given that the most recent post is 16 months old, we should consider
whether we even need the site any more.