Fix nav overlap issue on scroll
Before, when the nav became fixed, the page height would change and the
content would be hidden and shifted upwards. To solve this, the height
of the sticky nav is now taken into account and applied to the body
height as a margin. This ensures that the content is no longer hidden by
the nav on scroll, providing a better user experience.
These changes are only applied to screens larger than 768px, since the
original design is for the nav to not be sticky on smaller screens.