Fix flickering caused by insufficient body height
Previously, the height of the page would drastically reduce due to the
applied .sticky class on the global-header, causing the class to be
constantly removed and added on scroll. To resolve this issue, the body
has been set to occupy the full screen height and the main landmark to
grow and fill any empty space. These changes ensure that the page height
remains consistent.