Add d-marque.png back to fix a 404
In commit 08ccfb6bf0b18385e501a6406b30ce004161ed42
we added some assets for the new brand. However, the
d-marque image was added to the static directory, which
holds compiled assets, not the assets directory which
holds uncompiled assets.

Because of this, when we next rebuilt the theme in commit
a12a3f28b1c2414f50e9245ff36c90b982f99559 the 'clean' Grunt
jobs removed the logo and the site has been serving a 404
for that asset ever since.

This commit puts the image back in both the compiled and
uncompiled directories.