Refresh brand palette
The current implementation of the dxw brand palette is
quite jarring. The primary and accent colours are orange
and purple, which are roughly opposite each other on the
colour wheel, making the homepage in particular high
contrast and unattractive for long reading times.

The new brand was applied in 08ccfb6 (2022) and described
as "a very rough pass to apply the brand".

This commit chooses a split-complementary (triadic)
set of colours, retaining purple as the main brand colour.

In addition we use the teal and pink as accent colours,
forming a more balanced and elegant palette. For each
primary or accent colour we choose the lightest version,
which should be easier to view on web interfaces where
visitors may spend longer periods of time.

Practically, this commit changes, adds and removes just
enough of the palette to compile the theme. Choices
about how each element in the templates should be
styled will be made in later commits.