Add STS header for the site on non-local environments
Copied from the current Playbook configuration,
which doesn't have subdomains included.

Note that we ignore local environments here, which
are not expected to have SSL running. However, on
our own local Docker images, we set a 'production'
environment and do not (by default) have https.
Because we are not connecting to the site via https,
browsers ignore the STS header, and so this does
not cause problems in practice.