Add strict CSP which only allows analytics and wp.org
This site only contacts Plausible (for analytics) and
wordpress.org (for some core functionality).

Anything else can be disallowed, including Gravatar,
which means the icon on the top right of the admin
bar will not appear.

In some cases we've had to use unsafe-inline, partly
because some scripts in WordPress Core need this.
Not all of those scripts are needed though, for
example one is to check whether the customizer
is available.