FIX hover and active on search input
Add top and bottom margin so that highlight outline can be seen
all the way around the search input.

Ensure the outline can be seen when pressing the button, with
z-index and removing the margin on the right of the
search input.

Ensure the 1px offset bug does not reappear, shift the
margin directives into the container, use position
relative, equalise the line heights of both elements.