More consistent links and hover states
Previously there were several different sorts of hover
and active states for links, buttons and form input.

The following sorts of links all worked differently:

* inline links worked differently
* page titles on archive pages
* navigation bar links
* the pagination links at the bottom of post pages
* links in the footer
* the logo
* services tiles in the front page

This commit does not leave a perfect implementation of
links in different states, but it does leave a more
consistent implementation over the whole site.

Wherever possible, the code is more consistent by
removing CSS rather than adding it.

Links now have underlines by default, with the exception
of:

* the logo
* buttons
* service tiles and pagination links (underlines appear on hover)

On hover the underline is thickened and the colour of
the link is changed.

In the navigation bar there is still a thicker "underline"
over the page you are currently on, if you are on one of
the three pages in the nav. This has been adjusted to be
a little closer to the text so that the two different
text decorations do not seem to be competing with one
another.

In the footer we use solid underlines rather than dotted
decorations for untouched links, so colour contrast is
a little better.