Simplify pagination links
This commit simplifies the links we have for pagination,
firstly by removing assets (Javascript and images) that
we ship with this site just for pagination.

Secondly we use the_posts_pagination() now for creating
the links themselves, rather than the old h() call.

Lastly we simplify much of the styling, allowing most
of the styling to fall back on the base styles for
the site.