Simplify how archive lists are presented
Previously we've had a number of a11y issues on this site
due to headings being out of order. The way that archive
lists are styled is also different on the front page
and the two main archive pages and could be improved.

This is partly due to the archive lists being generated
in two different ways. On the archive pages we used
a partial and on the home page we did not.

This commit simplifies that a little, by using the
partial everywhere. The partial now needs a switch
to decide which heading level it should produce,
but this does mean we only have one piece of HTML
to style.