Skip to main content appears on first tab
This commit separates out the sr-only CSS class that
we already have into two classes, one of which is
specific to the skip link at the top of each page.

This allows us to treat that element differently to
other invisible elements, and to make it appear on
the screen on the first tab on any page in the site.